% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{empty_rgwrc_hold}
\alias{empty_rgwrc_hold}
\title{Empty rgwrc_hold}
\usage{
empty_rgwrc_hold()
}
\value{
Deletes all files in the rgwrc_hold directory.
}
\description{
Finds all the files in the rgwrc_hold directory and then deletes them.
}
\examples{
empty_rgwrc_hold();
}
