% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{euc.dist}
\alias{euc.dist}
\title{Euclidean distance in 2-D}
\usage{
euc.dist(x1, x2)
}
\arguments{
\item{x}{x coord}

\item{y}{y coord}
}
\value{
eucledian distance between a vector of 2-D points
}
\description{
Assign initial fixed equidistant centers for consistent Kmeans results
}
