% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{gwr.predict_py}
\alias{gwr.predict_py}
\title{GWR Predict}
\usage{
gwr.predict_py(
  salesdata = mgwrdata,
  populationdata = mgwrpopData,
  kernel = "gaussian",
  latlong = TRUE
)
}
\arguments{
\item{salesdata}{Model data.frame first 3 columns must be ordered coords then sales}

\item{populationdata}{Population data.frame first 2 columns must be ordered coords}

\item{kernel}{kernel distribution function: 'gaussian','bisquare','gaussian' is the default}
}
\value{
predicted value data.frame of the Model assigned
directly to the global environment, predicted value data.frame for
the population assigned directly to the global environment, and GWR summary results
}
\description{
Find the GWR predicted values for the population given some Model data.
Include an ID column in the table as the 4th column to preserve order.
}
\examples{
mgwrdata <- head(zillow_1k,300) # the model data with independent variable
mgwrpopData <- tail(subset(zillow_1k, select = -value),700)
gwr.predict_py(
  salesdata = mgwrdata,
  populationdata = mgwrpopData,
  kernel = "bisquare",
  latlong = FALSE
)
}
