% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{gwr_py}
\alias{gwr_py}
\title{GWR}
\usage{
gwr_py(salesdata = mgwrdata, kernel = "gaussian", latlong = TRUE)
}
\arguments{
\item{salesdata}{Model data.frame first 3 columns must be ordered coords then sales}

\item{kernel}{kernel distribution function: 'gaussian','bisquare','gaussian' is the default}
}
\value{
predicted values data.frame and GWR summary results
}
\description{
Find the GWR predicted values given some Model data
}
\examples{
mgwrdata <- zillow_1k

gwrresults <- gwr_py(salesdata = mgwrdata, kernel = "bisquare", latlong = FALSE)

}
