% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{initialKMeansStartingPoints}
\alias{initialKMeansStartingPoints}
\title{Initial Kmeans Starting Points}
\usage{
initialKMeansStartingPoints(x, n = 10)
}
\arguments{
\item{x}{coords}

\item{n}{number of clusters}
}
\value{
Fixed equidistant centers
}
\description{
Assign initial fixed equidistant centers for consistent Kmeans results
}
