% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{mgwr_c_bf_cpy}
\alias{mgwr_c_bf_cpy}
\title{Geographically Weighted Regression Clustering Best Fit Function}
\usage{
mgwr_c_bf_cpy(
  yf,
  x,
  data,
  direc,
  lonlat,
  weight,
  clustnum,
  Fixed_K = "Yes",
  bestC = "No",
  distmethod = "euclidean",
  hclustmethod = "complete"
)
}
\arguments{
\item{yf}{Dependent variable}

\item{x}{Independent variables}

\item{data}{data.frame}

\item{direc}{method to use for step wise regression options are 'none', 'both', 'forward', 'backward'}

\item{lonlat}{IF TRUE calculates the great circle distance, otherwise uses euclidean distance}

\item{weight}{kernel distribution function: 'gaussian','bisquare'}

\item{clustnum}{Number of clusters to run the function over}

\item{Fixed_K}{Assign initial starting centers for Kmeans for consistent results, for repeated analysis}

\item{bestC}{If "Yes" uses the NbClust package to determine the optimal number of clusters, can be time consuming, default is 'No'}

\item{distmethod}{The metric system to use for the distance matrix, default is 'euclidean'}

\item{hclustmethod}{Method to use for hclust function, default is 'complete'}
}
\value{
predicted values data.frame and GWR summary results
}
\description{
Assigns the weighted predictions from a GWR model in addition to coordinates
to a Kmeans and Hierarchical clustering algorithms. Regression Analysis is performed
on each cluster up to some designated max to determine what is the optimal neighborhood
to use for regression analysis. Polygon overlap of the clusters is taken into account.
}
\examples{
mgwr_c_bf_cpy(
yf = zillow_1k[,c("value"),drop = FALSE],
x = zillow_1k[,c("nbaths","nbeds","area","age")],
data = zillow_1k,
direc = 'backward',
lonlat = FALSE,
weight = 'gaussian',
clustnum = 8,
bestC = "No",
distmethod = "euclidean",
hclustmethod = "complete"
)
}
