% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{Mode}
\alias{Mode}
\title{Mode Function}
\usage{
Mode(x)
}
\arguments{
\item{x}{input vector}
}
\value{
Mode of the vector
}
\description{
Finds the most occurred feature in a vector
}
\examples{
x <- c(1,2,4,7,7,7,8,8,11,12,13,13,15);
Mode(x);
}
