% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{near.group.A}
\alias{near.group.A}
\title{Assign population clusters}
\usage{
near.group.A(data1, data2, crs_frame = 4326)
}
\arguments{
\item{data1}{Population data.frame containing the ID and coords}

\item{data2}{Model data.frame containing the ID, cluster and coords}

\item{crs_frame}{coordinate reference frame, default is Lat Long}
}
\value{
Mode of the vector
}
\description{
Assign each element of population their cluster based off the nearest
respective model data element cluster.
}
