% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PEcAn.yourpkg}
\alias{PEcAn.yourpkg}
\title{Geographically Weighted Regression Clustering}
\description{
Uses gwr and culster analysis for automated neighborhood generation,
in conjuction with geo-spatial analysis functions.
https://mgwr.readthedocs.io/en/latest/index.html
citations here etc.
}
