% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwrc.R
\name{rgwrc_miniconda_info}
\alias{rgwrc_miniconda_info}
\title{MiniConda Installation Info}
\usage{
rgwrc_miniconda_info()
}
\description{
Tells the user how to install miniconda and the python packages to use.
}
\examples{
rgwrc_miniconda_info();
}
