% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zillow_1k}
\alias{zillow_1k}
\title{price and details from zillow of 1,000 homes}
\format{
A data frame with 1000 rows and 7 variables:
\describe{
  \item{utmX}{X-coordinate}
  \item{utmY}{Y-coordinate}
  \item{value}{price}
  \item{nbath}{number of bathrooms}
  \item{nbed}{number of bedrooms}
  \item{area}{living area}
  \item{age}{home age}
}
}
\usage{
zillow_1k
}
\description{
A dataset containing the prices and other attributes of 1,000 homes.
The variables are as follows:
}
\keyword{datasets}
